% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.ukb-package.R
\docType{package}
\name{png.ukb-package}
\alias{png.ukb}
\alias{png.ukb-package}
\title{png.ukb: Exploring UKB Data}
\description{
Functions to load and wrangle UK Biobank data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/ukbwranglr/}
  \item \url{https://github.com/rmgpanw/ukbwranglr}
  \item Report bugs at \url{https://github.com/rmgpanw/ukbwranglr/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com}

}
\keyword{internal}
